Move StringSet to cloud/tasks/common
