+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/storage/storage.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ydb-platform/nbs/cloud/disk_manager/internal/pkg/dataplane/common"
 	"github.com/ydb-platform/nbs/cloud/disk_manager/internal/pkg/dataplane/snapshot/storage/protos"
 	"github.com/ydb-platform/nbs/cloud/disk_manager/internal/pkg/types"
-	task_storage "github.com/ydb-platform/nbs/cloud/tasks/storage"
+	tasks_common "github.com/ydb-platform/nbs/cloud/tasks/common"
 )
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -145,5 +145,5 @@ type Storage interface {
 		disk *types.Disk,
 	) (snapshotID string, checkpointID string, err error)
 
-	ListSnapshots(ctx context.Context) (task_storage.StringSet, error)
+	ListSnapshots(ctx context.Context) (tasks_common.StringSet, error)
 }
