+++ cloud/tasks/storage/storage_ydb_test.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	"github.com/ydb-platform/nbs/cloud/tasks/common"
 	tasks_config "github.com/ydb-platform/nbs/cloud/tasks/config"
 	"github.com/ydb-platform/nbs/cloud/tasks/errors"
 	"github.com/ydb-platform/nbs/cloud/tasks/logging"
@@ -102,7 +103,7 @@ func TestStorageYDBCreateTask(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	}
 
 	metricsRegistry.GetCounter(
@@ -142,7 +143,7 @@ func TestStorageYDBCreateTaskIgnoresID(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	}
 
 	metricsRegistry.GetCounter(
@@ -182,7 +183,7 @@ func TestStorageYDBCreateTwoTasks(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	}
 
 	metricsRegistry.GetCounter(
@@ -234,7 +235,7 @@ func TestStorageYDBCreateTwoTasksWithDifferentIdempotencyKeys(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	}
 
 	metricsRegistry.GetCounter(
@@ -287,7 +288,7 @@ func TestStorageYDBCreateTwoTasksWithSameIdempotencyKey(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	}
 
 	metricsRegistry.GetCounter(
@@ -334,7 +335,7 @@ func TestStorageYDBFailCreationOfTwoTasksWithSameIdempotencyKeyButDifferentTypes
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	}
 
 	metricsRegistry.GetCounter(
@@ -393,7 +394,7 @@ func TestStorageYDBGetTask(t *testing.T) {
 		GenerationID:   42,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{1, 2, 3},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		ZoneID:         "zone",
 	})
 	require.NoError(t, err)
@@ -409,7 +410,7 @@ func TestStorageYDBGetTask(t *testing.T) {
 	require.EqualValues(t, 42, taskState.GenerationID)
 	require.EqualValues(t, TaskStatusReadyToRun, taskState.Status)
 	require.EqualValues(t, []byte{1, 2, 3}, taskState.State)
-	require.EqualValues(t, NewStringSet(), taskState.Dependencies)
+	require.EqualValues(t, common.NewStringSet(), taskState.Dependencies)
 	require.WithinDuration(t, time.Time(createdAt), time.Time(taskState.ChangedStateAt), time.Microsecond)
 	require.EqualValues(t, "zone", taskState.ZoneID)
 	metricsRegistry.AssertAllExpectations(t)
@@ -453,7 +454,7 @@ func TestStorageYDBGetTaskWithDependencies(t *testing.T) {
 		GenerationID:   42,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -467,7 +468,7 @@ func TestStorageYDBGetTaskWithDependencies(t *testing.T) {
 		GenerationID:   42,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -481,7 +482,7 @@ func TestStorageYDBGetTaskWithDependencies(t *testing.T) {
 		GenerationID:   42,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{1, 2, 3},
-		Dependencies:   NewStringSet(depID1, depID2),
+		Dependencies:   common.NewStringSet(depID1, depID2),
 	})
 	require.NoError(t, err)
 
@@ -496,7 +497,7 @@ func TestStorageYDBGetTaskWithDependencies(t *testing.T) {
 	require.EqualValues(t, 42, taskState.GenerationID)
 	require.EqualValues(t, TaskStatusWaitingToRun, taskState.Status)
 	require.EqualValues(t, []byte{1, 2, 3}, taskState.State)
-	require.EqualValues(t, NewStringSet(depID1, depID2), taskState.Dependencies)
+	require.EqualValues(t, common.NewStringSet(depID1, depID2), taskState.Dependencies)
 	require.WithinDuration(t, time.Time(createdAt), time.Time(taskState.ChangedStateAt), time.Microsecond)
 	metricsRegistry.AssertAllExpectations(t)
 }
@@ -554,7 +555,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -568,7 +569,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -582,7 +583,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -596,7 +597,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -610,7 +611,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusFinished,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -624,7 +625,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -638,7 +639,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -652,7 +653,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -666,7 +667,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -680,7 +681,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelled,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -694,7 +695,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -708,7 +709,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -722,7 +723,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -736,7 +737,7 @@ func TestStorageYDBListTasksReadyToRun(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -794,7 +795,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -808,7 +809,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -822,7 +823,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -836,7 +837,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -850,7 +851,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusFinished,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -864,7 +865,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -878,7 +879,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -892,7 +893,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -906,7 +907,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -920,7 +921,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelled,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -934,7 +935,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -948,7 +949,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -962,7 +963,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -976,7 +977,7 @@ func TestStorageYDBListTasksReadyToCancel(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -1030,7 +1031,7 @@ func (f hangingTaskTestFixture) createTask(
 		GenerationID:   10,
 		Status:         taskStatus,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	}
 	if estimatedDuration > 0 {
 		state.EstimatedTime = createdAt.Add(estimatedDuration)
@@ -1235,7 +1236,7 @@ func TestStorageYDBListTasksRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1249,7 +1250,7 @@ func TestStorageYDBListTasksRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1263,7 +1264,7 @@ func TestStorageYDBListTasksRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1277,7 +1278,7 @@ func TestStorageYDBListTasksRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1291,7 +1292,7 @@ func TestStorageYDBListTasksRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusFinished,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1305,7 +1306,7 @@ func TestStorageYDBListTasksRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1319,7 +1320,7 @@ func TestStorageYDBListTasksRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1333,7 +1334,7 @@ func TestStorageYDBListTasksRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1347,7 +1348,7 @@ func TestStorageYDBListTasksRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1361,7 +1362,7 @@ func TestStorageYDBListTasksRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelled,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1413,7 +1414,7 @@ func TestStorageYDBListTasksCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1427,7 +1428,7 @@ func TestStorageYDBListTasksCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1441,7 +1442,7 @@ func TestStorageYDBListTasksCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1455,7 +1456,7 @@ func TestStorageYDBListTasksCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1469,7 +1470,7 @@ func TestStorageYDBListTasksCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusFinished,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1483,7 +1484,7 @@ func TestStorageYDBListTasksCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1497,7 +1498,7 @@ func TestStorageYDBListTasksCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1511,7 +1512,7 @@ func TestStorageYDBListTasksCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1525,7 +1526,7 @@ func TestStorageYDBListTasksCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1539,7 +1540,7 @@ func TestStorageYDBListTasksCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelled,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1596,7 +1597,7 @@ func TestStorageYDBListFailedTasks(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{0},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1610,7 +1611,7 @@ func TestStorageYDBListFailedTasks(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{0},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	_, err = storage.UpdateTask(ctx, TaskState{
@@ -1640,7 +1641,7 @@ func TestStorageYDBListFailedTasks(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{0},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	_, err = storage.UpdateTask(ctx, TaskState{
@@ -1709,7 +1710,7 @@ func TestStorageYDBListSlowTasks(t *testing.T) {
 			GenerationID:   generationID,
 			Status:         TaskStatusFinished,
 			State:          []byte{0},
-			Dependencies:   NewStringSet(),
+			Dependencies:   common.NewStringSet(),
 			EndedAt:        created.Add(time.Duration(durationMinutes) * time.Minute),
 			EstimatedTime:  created.Add(time.Duration(estimated) * time.Minute),
 		}
@@ -1798,7 +1799,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1812,7 +1813,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1826,7 +1827,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1840,7 +1841,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastHost:       "other_host",
 	})
 	require.NoError(t, err)
@@ -1855,7 +1856,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusFinished,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1869,7 +1870,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1883,7 +1884,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1897,7 +1898,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1911,7 +1912,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1925,7 +1926,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelled,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -1939,7 +1940,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -1953,7 +1954,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -1967,7 +1968,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -1981,7 +1982,7 @@ func TestStorageYDBListTasksStallingWhileRunning(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -2039,7 +2040,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2053,7 +2054,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2067,7 +2068,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2081,7 +2082,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2095,7 +2096,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusFinished,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2109,7 +2110,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2123,7 +2124,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2137,7 +2138,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2151,7 +2152,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastHost:       "other_host",
 	})
 	require.NoError(t, err)
@@ -2166,7 +2167,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelled,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2180,7 +2181,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -2194,7 +2195,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -2208,7 +2209,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToRun, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -2222,7 +2223,7 @@ func TestStorageYDBListTasksStallingWhileCancelling(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(taskIDFinished, taskIDReadyToCancel),
+		Dependencies:   common.NewStringSet(taskIDFinished, taskIDReadyToCancel),
 	})
 	require.NoError(t, err)
 
@@ -2279,7 +2280,7 @@ func TestStorageYDBListTasksReadyToRunWithWhitelist(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2293,7 +2294,7 @@ func TestStorageYDBListTasksReadyToRunWithWhitelist(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2343,7 +2344,7 @@ func TestStorageYDBListTasksReadyToCancelWithWhitelist(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2357,7 +2358,7 @@ func TestStorageYDBListTasksReadyToCancelWithWhitelist(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -2407,7 +2408,7 @@ func TestStorageYDBListTasksStallingWhileRunningWithWhitelist(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastHost:       "other_host",
 	})
 	require.NoError(t, err)
@@ -2422,7 +2423,7 @@ func TestStorageYDBListTasksStallingWhileRunningWithWhitelist(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastHost:       "other_host",
 	})
 	require.NoError(t, err)
@@ -2474,7 +2475,7 @@ func TestStorageYDBListTasksStallingWhileCancellingWithWhitelist(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastHost:       "other_host",
 	})
 	require.NoError(t, err)
@@ -2489,7 +2490,7 @@ func TestStorageYDBListTasksStallingWhileCancellingWithWhitelist(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastHost:       "other_host",
 	})
 	require.NoError(t, err)
@@ -2543,7 +2544,7 @@ func TestStorageYDBListTasksReadyToRunInCertainZone(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		ZoneID:         "zone0",
 	})
 	require.NoError(t, err)
@@ -2558,7 +2559,7 @@ func TestStorageYDBListTasksReadyToRunInCertainZone(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		ZoneID:         "zone1",
 	})
 	require.NoError(t, err)
@@ -2607,7 +2608,7 @@ func TestStorageYDBListTasksReadyToCancelInCertainZone(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		ZoneID:         "zone0",
 	})
 	require.NoError(t, err)
@@ -2622,7 +2623,7 @@ func TestStorageYDBListTasksReadyToCancelInCertainZone(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		ZoneID:         "zone1",
 	})
 	require.NoError(t, err)
@@ -2671,7 +2672,7 @@ func TestStorageYDBListTasksStallingWhileRunningInCertainZone(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastHost:       "other_host",
 		ZoneID:         "zone0",
 	})
@@ -2687,7 +2688,7 @@ func TestStorageYDBListTasksStallingWhileRunningInCertainZone(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastHost:       "other_host",
 		ZoneID:         "zone1",
 	})
@@ -2742,7 +2743,7 @@ func TestStorageYDBListTasksStallingWhileCancellingInCertainZone(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastHost:       "other_host",
 		ZoneID:         "zone0",
 	})
@@ -2758,7 +2759,7 @@ func TestStorageYDBListTasksStallingWhileCancellingInCertainZone(t *testing.T) {
 		GenerationID:   generationID,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastHost:       "other_host",
 		ZoneID:         "zone1",
 	})
@@ -2815,7 +2816,7 @@ func TestStorageYDBLockTaskToRun(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastRunner:     "runner_42",
 	})
 	require.NoError(t, err)
@@ -2868,7 +2869,7 @@ func TestStorageYDBLockTaskToRunWrongGeneration(t *testing.T) {
 		GenerationID:   1,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -2914,7 +2915,7 @@ func TestStorageYDBLockTaskToRunWrongState(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -2961,7 +2962,7 @@ func TestStorageYDBLockTaskToCancel(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToCancel,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastRunner:     "runner_42",
 	})
 	require.NoError(t, err)
@@ -3014,7 +3015,7 @@ func TestStorageYDBLockTaskToCancelWrongGeneration(t *testing.T) {
 		GenerationID:   1,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3060,7 +3061,7 @@ func TestStorageYDBLockTaskToCancelWrongState(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3107,7 +3108,7 @@ func TestStorageYDBMarkForCancellation(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastRunner:     "runner_42",
 	})
 	require.NoError(t, err)
@@ -3159,7 +3160,7 @@ func TestStorageYDBMarkForCancellationIfAlreadyCancelling(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3209,7 +3210,7 @@ func TestStorageYDBMarkForCancellationIfAlreadyFinished(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusFinished,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3258,7 +3259,7 @@ func TestStorageYDBMarkForCancellationIfAlreadyCancelled(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusCancelled,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3307,7 +3308,7 @@ func TestStorageYDBUpdateTask(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{0},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -3321,7 +3322,7 @@ func TestStorageYDBUpdateTask(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -3335,7 +3336,7 @@ func TestStorageYDBUpdateTask(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusFinished,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3353,7 +3354,7 @@ func TestStorageYDBUpdateTask(t *testing.T) {
 		ErrorCode:      grpc_codes.InvalidArgument,
 		ErrorMessage:   "invalid argument",
 		State:          []byte{1},
-		Dependencies:   NewStringSet(taskIDDependent1, taskIDDependent2),
+		Dependencies:   common.NewStringSet(taskIDDependent1, taskIDDependent2),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3384,7 +3385,7 @@ func TestStorageYDBUpdateTask(t *testing.T) {
 		ErrorCode:      grpc_codes.InvalidArgument,
 		ErrorMessage:   "invalid argument",
 		State:          []byte{1},
-		Dependencies:   NewStringSet(taskIDDependent1, taskIDDependent2),
+		Dependencies:   common.NewStringSet(taskIDDependent1, taskIDDependent2),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3436,7 +3437,7 @@ func TestStorageYDBUpdateTaskStatus(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{0},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3452,7 +3453,7 @@ func TestStorageYDBUpdateTaskStatus(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{1},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastRunner:     "runner_42",
 	})
 	require.NoError(t, err)
@@ -3480,7 +3481,7 @@ func TestStorageYDBUpdateTaskStatus(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusRunning,
 		State:          []byte{2},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3524,7 +3525,7 @@ func TestStorageYDBUpdateTaskWrongGeneration(t *testing.T) {
 		GenerationID:   2,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3540,7 +3541,7 @@ func TestStorageYDBUpdateTaskWrongGeneration(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusRunning,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.Truef(t, errors.Is(err, errors.NewWrongGenerationError()), "Expected WrongGenerationError got %v", err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3584,7 +3585,7 @@ func TestStorageYDBLockInParallel(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastRunner:     "runner_42",
 	})
 	require.NoError(t, err)
@@ -3649,7 +3650,7 @@ func TestStorageYDBMarkForCancellationInParallel(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastRunner:     "runner_42",
 	})
 	require.NoError(t, err)
@@ -3706,7 +3707,7 @@ func TestStorageYDBLockAlreadyCancellingTask(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 	metricsRegistry.AssertAllExpectations(t)
@@ -3758,7 +3759,7 @@ func TestStorageYDBCheckStallingTimeout(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusCancelling,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastHost:       "other_host",
 	})
 	require.NoError(t, err)
@@ -4418,7 +4419,7 @@ func TestStorageYDBClearEndedTasks(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{0},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -4432,7 +4433,7 @@ func TestStorageYDBClearEndedTasks(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	}
 	taskID2, err := storage.CreateTask(ctx, task)
 	require.NoError(t, err)
@@ -4452,7 +4453,7 @@ func TestStorageYDBClearEndedTasks(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	}
 
 	taskID3, err := storage.CreateTask(ctx, task)
@@ -4476,7 +4477,7 @@ func TestStorageYDBClearEndedTasks(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	}
 
 	taskID4, err := storage.CreateTask(ctx, task)
@@ -4539,7 +4540,7 @@ func TestStorageYDBPauseResumeTask(t *testing.T) {
 		GenerationID:   0,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 		LastRunner:     "runner_42",
 	})
 	require.NoError(t, err)
@@ -4735,7 +4736,7 @@ func TestForceFinishTaskWithDependencies(t *testing.T) {
 		GenerationID:   42,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -4749,7 +4750,7 @@ func TestForceFinishTaskWithDependencies(t *testing.T) {
 		GenerationID:   42,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{},
-		Dependencies:   NewStringSet(),
+		Dependencies:   common.NewStringSet(),
 	})
 	require.NoError(t, err)
 
@@ -4763,7 +4764,7 @@ func TestForceFinishTaskWithDependencies(t *testing.T) {
 		GenerationID:   42,
 		Status:         TaskStatusReadyToRun,
 		State:          []byte{1, 2, 3},
-		Dependencies:   NewStringSet(depID1, depID2),
+		Dependencies:   common.NewStringSet(depID1, depID2),
 	})
 	require.NoError(t, err)
